Fix a bug in batch repayment window
The algorithm that fetches due amounts in the batch repayment window
(available via View -> Search -> Double-click) was built on the
assumption that the date of today is _before_ the last expected date,
which is wrong, obviously. And thus if you tried to repay a loan whose
last repayment date is in the past, you would get an exception.

OP-65